fixed Pagination type + moved interfaces to dedicated file
