+++ services/metadata/frontend/content/Content.svelte
@@ -2,13 +2,13 @@
 import Tile from "./Tile.svelte";
 import Category from "./Category.svelte";
 import { onMount } from "svelte";
-import type { Project } from "./project.model";
+import type { PaginationData, Project } from "./interfaces";
 import Pagination from "./Pagination.svelte";
 import { getProjects, pagedResults, pages } from "./stores";
 
 let projects: Project[];
 let message = 'Loading...';
-let pagination: object;
+let pagination: PaginationData;
 
 setTimeout(() => {
   const noData = 'No data retrived. Please check the connection and retry.';
@@ -19,11 +19,9 @@ setTimeout(() => {
 onMount(async () => {
   await getProjects(1);
 
-  pagedResults.subscribe(r => {
-    projects = r;
-  });
+  pagedResults.subscribe(r => projects = r);
 
-  pages.subscribe(t => pagination = t);
+  pages.subscribe(p => pagination = p);
 });
 </script>
