--- services/metadata/frontend/content/project.model.ts
@@ -1,5 +1,0 @@
-export interface Project {
-  description: string;
-  id: string;
-  name: string;
-}
